[Refactor] Move and restructure Servers components
