+++ src/pages/Servers/Servers.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Header } from "@/components/sections/Header";
-import Fetching from "@/components/layouts/Servers/Table/Fetching";
-import YourServer from "@/components/layouts/Servers/YourServer/YourServer";
+import Fetching from "@/components/layouts/Servers/Fetching/Fetching";
+import YourServer from "@/components/layouts/Servers/YourServer";
 import { Footer } from "@/components/sections/Footer";
 
 const Servers = () => {
